srs/requirements.txt
Django==2.2.7
django-cors-headers==3.2.0
django-countries==5.5
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
Pillow==6.2.1
pkg-resources==0.0.0
PyJWT==1.7.1
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
